Consolidate service naming and add barrel exports
