+++ src/services/index.ts
@@ -1,0 +1,13 @@
+// Service exports for clean imports
+
+// Class-based services (extracted from MessageDB - need instantiation)
+export { MessageService } from './MessageService';
+export { EncryptionService } from './EncryptionService';
+export { SpaceService } from './SpaceService';
+export { SyncService } from './SyncService';
+export { ConfigService } from './ConfigService';
+export { InvitationService } from './InvitationService';
+
+// Utility services
+export { SearchService } from './SearchService';
+export { notificationService } from './NotificationService';
